backend/src/requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-cors-headers==3.7.0
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
kombu==5.1.0
numpy==1.20.3
pandas==1.2.4
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
ta==0.7.0
vine==5.0.0
wcwidth==0.2.5
